Mock proofs and variable vk deployment. (#669)
* Consolidate vk construction scripts in prep for offloading to ci. Make possible to deploy diff vk in diff contexts.

* Proverless mode is now really 'mock proof' mode. Speeds up bootstrap. Also uses real proofs, verifier etc.

* Mainframe updates.

* Fix bb tests.

* Big refactor rollup_cli verify.cpp/hpp files to reduce duplication and improve logging. Fixes.

* Target skylake again as build being slow. Disable test discovery in CI as we don't use it anyway.

* Intial remote script run generalisation.

* Initial remote build.

* All bb builds on aws.

* Split up root rollup full tests.

* Build blockchain on 64 core aws. Build vks. We still don't build real 28x32 as it would still take too long. This will add ~4.5 mins to build time, but will simplify our dev depolyments.

* Falafel proverless no longer skips validation.

* Increase spot price.

+++ blockchain/src/contracts/verifier/fixtures/create_fixture.sh
@@ -1,5 +1,12 @@
-#!/bin/sh
+#!/bin/bash
 set -e
+
+MOCK_PROOF=$1
+CONDITIONAL=$2
+
+if [ -n "$CONDITIONAL" -a -f rollup_proof_data_3x2.dat ]; then
+  exit 0;
+fi
 
 cd ../../../../../barretenberg
 mkdir -p build && cd build && rm -rf ./data && cmake .. && make -j$(nproc) tx_factory rollup_cli
@@ -7,4 +14,4 @@ cd ../srs_db && ./download_ignition.sh 1
 cd ../../blockchain/src/contracts/verifier/fixtures
 
 # input format: num_txs | data_dir | inner size | outer size | valid outer sizes
-./create_rollup_proof.sh 4 ./data 3 2
+./create_rollup_proof.sh 4 ./data 3 2 $MOCK_PROOF
